Rename ws migrate to migrate-definition (#1598)
