+++ src/snowflake/cli/_plugins/workspace/commands.py
@@ -20,7 +20,6 @@ from textwrap import dedent
 from typing import List, Optional
 
 import typer
-import yaml
 from snowflake.cli._plugins.nativeapp.artifacts import BundleMap
 from snowflake.cli._plugins.nativeapp.common_flags import (
     ForceOption,
@@ -34,44 +33,13 @@ from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
 from snowflake.cli.api.entities.common import EntityActions
 from snowflake.cli.api.exceptions import IncompatibleParametersError
 from snowflake.cli.api.output.types import MessageResult, QueryResult
-from snowflake.cli.api.project.definition_conversion import (
-    convert_project_definition_to_v2,
-)
-from snowflake.cli.api.project.definition_manager import DefinitionManager
-from snowflake.cli.api.secure_path import SecurePath
 
 ws = SnowTyperFactory(
     name="ws",
     help="Deploy and interact with snowflake.yml-based entities.",
+    is_hidden=lambda: True,
 )
 log = logging.getLogger(__name__)
-
-
-@ws.command()
-def migrate(
-    accept_templates: bool = typer.Option(
-        False, "-t", "--accept-templates", help="Allows the migration of templates."
-    ),
-    **options,
-):
-    """Migrates the Snowpark, Streamlit, and Native App project definition files from V1 to V2."""
-    manager = DefinitionManager()
-    pd = manager.unrendered_project_definition
-
-    if pd.meets_version_requirement("2"):
-        return MessageResult("Project definition is already at version 2.")
-
-    pd_v2 = convert_project_definition_to_v2(manager.project_root, pd, accept_templates)
-
-    SecurePath("snowflake.yml").rename("snowflake_V1.yml")
-    with open("snowflake.yml", "w") as file:
-        yaml.dump(
-            pd_v2.model_dump(
-                exclude_unset=True, exclude_none=True, mode="json", by_alias=True
-            ),
-            file,
-        )
-    return MessageResult("Project definition migrated to version 2.")
 
 
 @ws.command(requires_connection=True, hidden=True)
